Use Inconsolata font for urxvt and Vim
Use Inconsolata font for rxvt-unicode. Set antialias=true and
hinting=slight in the configuration for this font. Without
hinting, the font looks a little bit dull. The hinting provides
a slightly better contrast.

For Vim, use Inconsolata for both Linux as well as Windows.
However, add fallback fonts for both systems in case Inconsolata
is unavailable. The fallback font chosen for Linux is DejaVu
Sans Mono, although this fallback font is never used for Linux.
The fc-match tool always returns DejaVu Sans as the default font
when the preferred font is not found. The fallback font for
Windows is chosen as Consolas. Vim indeed falls back to Consolas
when Inconsolata is unavailable.

Note: On Windows, the Inconsolata font installed using the OTF
file from http://levien.com/type/myfonts/inconsolata.html
results in a font that appears a little rough and short.

The solution to the above issue is to install the font using the
TTF file from
http://googlefontdirectory.googlecode.com/hg/ofl/inconsolata/Inconsolata-Regular.ttf
because this results in a smooth and a little bit taller looking
font.