Reduce the font size used in gVim for Windows
Reduce the font size of Inconsolata from h18 to h16 and the font
size of the fallback font Consolas from h16 to h14 in gVim.

With this reduced font size, longer lines of code fit
horizontally without wrapping while working with vertical split
windows.

Here are the details of the resolution of both monitors, the
total number of columns that fit in gVim when maximized in both
monitors, and the total number of columns per window when there
are two vertical split windows equally wide.

Lenovo ThinkPad T420s 14" built-in display:
  - Resolution: 1600 x 900
  - Number of columns for gVim: 159
  - Number of columns per window: 79

Dell 1908 FP 19" monitor:
  - Resolution: 1280 x 1024
  - Number of columns for gVim: 127
  - Number of columns per window: 63