Reduce Vim and urxvt font size on Linux to 14pt
My Acer Extensa 4620 14" built-in display has a resolution of
1280 x 800. With 14pt Inconsolata, it is able to display 141
columns in Vim, gVim and rxvt-unicode. As a result, when there
are two vertical split windows in Vim, both equally wide, each
window displays 70 columns.

While this is not good enough, this is a good compromise between
the less number of columns (127 columns) available with 15pt and
the reduced readability with 13pt.