Reduce urxvt font size from 18px to 16px
In commit 5b0d95d4f4b6ec33edb36aebfb1ee1ec088f8f04, the font size of
urxvt was set to 18px. This allowed only 141 characters to fit in a
line. This is not good enough while working on Vim with vertical splits.
Therefore, reduce the font size to 16px which allows 159 characters to
fit in a line.