Set Monospace as fallback font for urxvt
With Inconsolata, some Unicode characters are not displayed. For
example, with Inconsolata, the euro sign is rendered correctly.

    $ printf "\xe2\x99\xa5\n"
    ♥

However, the black heart suit is not rendered correctly. It
appears as a box because Inconsolata does not support this
character.

    $ printf "\xe2\x82\xac\n"
    €

In order to support such characters, set Monospace as a fallback
font for urxvt. On my Debian 8.0 (Jessie) system, Monospace
matches DejaVu Sans Mono.

    $ fc-match Monospace
    DejaVuSansMono.ttf: "DejaVu Sans Mono" "Book"

Monospace is set as fallback font by appending it to Urxvt.font
attribute in ~/.Xresources.