Create a symbolic link ~/bin/vi.sh
I have a private todo management script that launches the todo file in
an editor. The script tries to launch the editor by executing the
following commands: vi.cmd, vi.sh, mvim, gvim, vi.

The vi.cmd and vi.sh scripts are meant to be identical to the Vimer
script. Therefore, add a symbolic link to the Vimer script at
~/bin/vi.sh so that the todo management script can launch the Vimer
script.