Perform upward search for tags file
The tags file is usually located in the top-level directory of a
project. However while editing source code, the current directory may be
one of the subdirectories nested within the top-level directory of the
project. Therefore, to ensure that the tag search commands work, set the
tags option such that the editor searches upwards for the tags file
starting with the current directory.

See :help file-searching (point 2) for more details on the syntax.