Change the beer prompt dynamically based on current day
This is an improvement upon the previous commit. The previous commit
sets the beer prompt only once during shell startup. If the day changes
while the shell session is running, the new prompts does not change
based on the current day, e.g., if a shell is created on Friday, the
beer prompt appears, and it remains so in the that session even on
Monday.

This commit ensures that every time a new prompt is displayed, it is
chosen based on the current day, e.g., if a shell is created on Friday,
the beer prompt appears, but on Monday, after running a command, the
dollar prompt appears again.

Additionally, on Saturday double-beer prompt appears.