Add Git aliases: loga, logg, and diffc
With these aliases:

  - `git loga` shows commit history of all branches,
  - `git logg` shows a graphical representation of the commit history of
     all branches, and
  - `git diffc` shows changes staged for the next commit.

In this commit, Git aliases for `git log` with `--decorate` are being
removed in Mac because the `--decorate` option is unnecessary since Git
2.13 because `--decorate` option is enabled by default now[1] and the
Git version on the current macOS High Sierra 10.3.3 installed with
Homebrew is 2.15.1.

The `--decorate` aliases are still necessary in Debian and CentOS though
because Debian 9 (stretch) has Git 2.11.0 and CentOS 7.4 has Git 1.8.3.1.

[1]: https://stackoverflow.com/a/25097285/303363