Bind tmux kill keys without confirm-before
The default tmux key bindings to kill windows and panes, `&` and `x`,
respectively, are mapped to the following commands:

    confirm-before -p "kill-window #W? (y/n)" kill-window
    confirm-before -p "kill-pane #P? (y/n)" kill-pane

As a result, when these key bindings are used, a confirmation prompt
appears before killing the window or pane. This can be inconvenient when
we want to kill a window or pane instantly without any prompts.

Therefore bind these keys with the kill commands directly without using
any `confirm-before` commands.