Set package-archives before package-initialize
Prior to this change, when a new package is added to the list of
packages in the package installation code (the `package-list`
variable), the installation of the new package fails with an error.
For example, if the `package-list` variable is updated to add a new
package named `markdown-mode` and then Emacs is restarted, the
following error occurs:

    error: Package ‘markdown-mode-’ is unavailable

This error occurs because prior to this fix, the Emacs Lisp code calls
`package-initialize` before adding MELPA to `package-archives`. As a
result, `package-initialize` is unaware of MELPA when it is called.
GNU ELPA is present in `package-archives` by default, so it attempts
to add only packages from GNU ELPA to the `package-archive-contents`
list when it performs the initialization tasks.

When GNU ELPA archive already exists at `~/.emacs.d/elpa/archives` due
to a previous invocation of the initialization file,
`package-initialize` succeeds in adding GNU ELPA archive contents to
`package-archive-contents`. As a result, the
`package-refresh-contents` call is skipped due to the surrounding
`unless package-archive-contents` expression. Thus MELPA archive
contents do not get added to `package-archive-contents` even if they
exist locally at `~/.emacs.d/elpa/archives`. Therefore
`package-install` fails to install a new package from MELPA and
reports the package as unavailable.

This change fixes this issue by ensuring that `package-initialize` is
called *after* adding MELPA to `package-archives`.