Turn em command into a shell function again
When the `em` command is moved out to a script, running `em` for the
first time starts Emacs with the home directory set as the current
working directory. Therefore, a command like `em foo.txt` always opens
`~/foo.txt` regardless of which directory the command is executed
from.

This issue can be demonstrated easily by running `sh -c emacs &` or
`sh -c nohup emacs &` and typing `C-h v default-directory` to see that
the value of this variable is the home directory.

However, running `emacs &` or `nohup emacs &` directly in the current
shell does not cause this problem. This is why the `em` command is now
changed to a shell function again.