Do not inherit fixed-pitch in org-table face
On macOS, the org-table face inherits fixed-pitch which uses the
Monospace family. Courier is used as a fallback for Monospace while
displaying the tables. This causes the tables to appear smaller than
the surrounding text.

This change prevents org-table from inheriting from fixed-pitch, so
that org-table text looks just like other text.