Set org-agenda-files after org loads
Prior to this change, opening an Org file and typing C-c C-w to refile
an entry led to the following error:

    No refile targets

This error occurred because prior to this change, the variable
org-agenda-files was being set within with-eval-after-load for
org-agenda.  So when C-c C-w was typed, although org and org-refile
were getting loaded, org-agenda was not getting loaded.  As a result,
org-agenda-files was not getting defined.

This change fixes this issue by setting org-agenda-files withi
with-eval-after-load for org instead.