Fix org-capture error due to nil buffer-file-name
The function enable-planner-settings which is added to org-mode-hook
performs string matching on buffer-file-name as follows:

    (string-match-p "/plan/.*\\.org$" buffer-file-name)

Now while running org-capture (C-c c), the buffer CAPTURE-in.org has
buffer-file-name set to nil.  As a result the above function call
fails with an error.  The following error is displayed:

    org-capture: Capture abort: Wrong type argument: stringp, nil

This change fixes this issue by ensuring that buffer-file-name is not
nil before performing the above function call.